Followup r7084
Not perfect (needs a bit of tweaking), but we can do page existence by checking the PageInfo cached...

Also, I suppose a category could exist if it's got members, but no text, or a blank page (this fixes that issue).. No idea what the API returns for non existant pages and no members, nothing maybe?

Bit of a todo left here.